refactor(api): client generation
