+++ packages/farrow-api-server/__tests__/expected/02.ts
@@ -4,10 +4,7 @@
  * Don't modify it manually
  */
 
-import { createApiPipelineWithUrl, ApiInvokeOptions } from 'farrow-api-client'
-
-import type { JsonType } from 'farrow-api-client'
-
+import type { JsonType } from 'farrow-api'
 /**
  * {@label Collection}
  */
@@ -94,15 +91,4 @@ export type PartialStruct = {
   a?: number | null | undefined
   b?: number | null | undefined
   c?: boolean | null | undefined
-}
-
-export const url = ''
-
-export const apiPipeline = createApiPipelineWithUrl(url)
-
-export const api = {
-  methodA: (input: Collection, options?: ApiInvokeOptions) =>
-    apiPipeline.invoke({ type: 'Single', path: ['methodA'], input }, options) as Promise<Collection>,
-  methodB: (input: Collection, options?: ApiInvokeOptions) =>
-    apiPipeline.invoke({ type: 'Single', path: ['methodB'], input }, options) as Promise<Collection>,
 }
